backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic[email]
motor
python-decouple
passlib